[UNDERTOW-466] Split out the generic notification handling side of the SecurityContext implementation into an abstract base class so future SecurityContext implementations can focus just on the actual authentication.
